% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_core.R
\name{plot_before_after_prevalence}
\alias{plot_before_after_prevalence}
\title{Plot Before/After Prevalence}
\usage{
plot_before_after_prevalence(data)
}
\arguments{
\item{data}{A QC data frame with \code{flag_exclude_authorized}.}
}
\value{
A ggplot object.
}
\description{
Plot Before/After Prevalence
}
