% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_core.R
\name{plot_prevalence_residuals}
\alias{plot_prevalence_residuals}
\title{Plot Prevalence Residuals}
\usage{
plot_prevalence_residuals(data)
}
\arguments{
\item{data}{A QC data frame with \code{pearson_resid} and \code{month_date}.}
}
\value{
A ggplot object.
}
\description{
Plot Prevalence Residuals
}
