% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_core.R
\name{plot_qc_summary_by_region}
\alias{plot_qc_summary_by_region}
\title{Plot QC Summary By Region}
\usage{
plot_qc_summary_by_region(summary_data)
}
\arguments{
\item{summary_data}{Output from \code{\link[=summarise_qc_by_region]{summarise_qc_by_region()}}.}
}
\value{
A ggplot object.
}
\description{
Plot QC Summary By Region
}
