% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_core.R
\name{plot_tested_robust_z}
\alias{plot_tested_robust_z}
\title{Plot Tested Robust Z}
\usage{
plot_tested_robust_z(data)
}
\arguments{
\item{data}{A QC data frame with \code{tested_robust_z} and \code{month_date}.}
}
\value{
A ggplot object.
}
\description{
Plot Tested Robust Z
}
