% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_core.R
\docType{package}
\name{routineqc-package}
\alias{routineqc}
\alias{routineqc-package}
\title{routineqc: Reusable QC Tools For Routine Surveillance Data}
\description{
The package provides modular functions for preparing routine facility-month
data, fitting prevalence models, flagging potential anomalies, and producing
summaries and plots for QC review.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jt-hicks/routineqc}
  \item Report bugs at \url{https://github.com/jt-hicks/routineqc/issues}
}

}
\author{
\strong{Maintainer}: Joseph T Hicks \email{jthicks@imperial.ac.uk}

}
\keyword{internal}
